+++ crates/wp-oml/src/core/evaluator/extract/basic/take.rs
@@ -3,7 +3,6 @@ use crate::core::prelude::*;
 use crate::language::EvaluationTarget;
 use crate::language::FieldTake;
 use wildmatch::WildMatch;
-use wp_model_core::model::data::record::RecordItem;
 use wp_model_core::model::{DataField, DataRecord};
 
 impl FieldExtractor for FieldTake {