+++ crates/wp-oml/src/parser/oml_conf.rs
@@ -13,7 +13,6 @@ use winnow::error::{ContextError, ErrMode, StrContext, StrContextValue};
 use winnow::stream::Stream;
 use wp_data_model::cache::FieldQueryCache;
 use wp_error::{OMLCodeError, OMLCodeResult};
-use wp_model_core::model::data::record::RecordItem;
 use wp_model_core::model::{DataField, DataRecord};
 use wp_parser::Parser;
 use wp_parser::WResult;
@@ -487,7 +486,6 @@ pub fn oml_conf_enable(data: &mut &str) -> WResult<bool> {
 mod tests {
     use crate::parser::oml_conf::oml_parse_raw;
     use crate::parser::utils::for_test::{assert_oml_parse, assert_oml_parse_ext};
-    use wp_model_core::model::data::record::RecordItem;
     use wp_parser::Parser;
     use wp_parser::WResult as ModalResult;
     use wp_parser::comment::CommentParser;